Replace ToReadOnlyCollection with AsReadOnlyList.
The semantics of ToReadOnlyCollection were confusing to many developers, as it doesn't guarantee that the underlying collection can't be mutated. (It was an "antipattern" to use it in a constructor and store the returned value in a field.)

The new AsReadOnlyList method provides the same efficiency, with a name that implies it's just a view, not a conversion.